finish implementation and fix all tests

+++ libs/commons/src/main/java/com/recom/commons/calculator/d8algorithm/D8AlgorithmForForestMap.java
@@ -49,8 +49,8 @@ public class D8AlgorithmForForestMap {
         final int forestCellSizeSquared = forestCellSizeInMeter * forestCellSizeInMeter;
         double forestDensityThreshold = 1F / 100; // @TODO extract to conf
 
-        final int spacialX = demX * demDescriptor.getStepSize();
-        final int spacialY = demY * demDescriptor.getStepSize();
+        final double spacialX = demX * demDescriptor.getStepSize().doubleValue();
+        final double spacialY = demY * demDescriptor.getStepSize().doubleValue();
 
         final List<ForestItem> forestItemsInSpace = spacialIndex.getInSpace(spacialX, spacialY);
         final double treeDensity = forestItemsInSpace.size() / (double) forestCellSizeSquared;
